Second compile fix to r24826.
TBR=bmeurer@chromium.org

Review URL: https://codereview.chromium.org/645163009

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24829 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/serialize.h
@@ -667,7 +667,7 @@ class StartupSerializer : public Serializer {
   }
 
  private:
-  int root_index_wave_front_;
+  intptr_t root_index_wave_front_;
   DISALLOW_COPY_AND_ASSIGN(StartupSerializer);
 };
